apiProject/requirement.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
pytest-django==4.4.0
model-bakery==1.3.2